requirements/requirements.txt
django==2.2.10
Werkzeug==1.0.1
ipip-ipdb==1.3.2
coreapi==2.3.3
pyotp==2.3.0
passlib==1.7.2
paramiko==2.7.1
sshpubkeys==3.1.0
itsdangerous==1.1.0
Pillow==7.1.1
celery==4.4.2
PyYAML==5.3.1
channels-redis==2.4.2
django-auth-ldap==2.1.1
django-celery-beat==2.0.0
django-fernet-fields==0.6
django-cors-headers==3.2.1
django-simple-captcha==0.5.12
django-redis-cache==2.1.0
django-rest-swagger==2.2.0
django-radius==1.5.0
django-filter==2.2.0
django-formtools==2.2
django-bootstrap4==1.1.1
django-amazon-ses==3.0.2
djangorestframework==3.10.3
djangorestframework-bulk==0.2.1
python-keycloak==0.13.3
python-keycloak-client==0.1.3
unicodecsv==0.14.1
aliyunsdkcore==1.0.3
aliyun-python-sdk-core-v3==2.13.11
httpsig==1.3.0
drf-yasg==1.17.1
psycopg2-binary==2.8.5
python-ldap==3.2.0
python-daemon==2.2.4